# crates/twerk-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Twerk API",
    "description": "Task scheduling and job queue management API",
    "contact": {
      "name": "Twerk Contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/triggers": {
      "get": {
        "tags": [
          "super::trigger_api::handlers::query"
        ],
        "summary": "GET /api/v1/triggers",
        "operationId": "list_triggers_handler",
        "responses": {
          "200": {
            "description": "List of triggers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TriggerView"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Persistence error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "super::trigger_api::handlers::command"
        ],
        "summary": "POST /api/v1/triggers",
        "operationId": "create_trigger_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Trigger created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerView"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/triggers/{id}": {
      "get": {
        "tags": [
          "super::trigger_api::handlers::query"
        ],
        "summary": "GET /api/v1/triggers/{id}",
        "operationId": "get_trigger_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Trigger ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trigger not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "super::trigger_api::handlers::command"
        ],
        "summary": "Handler for `PUT /api/v1/triggers/{id}`.",
        "operationId": "update_trigger_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Trigger ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Trigger updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerView"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trigger not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "super::trigger_api::handlers::query"
        ],
        "operationId": "delete_trigger_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Trigger ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Trigger deleted"
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trigger not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TriggerErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "body_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TriggerUpdateRequest": {
        "type": "object",
        "required": [
          "name",
          "enabled",
          "event",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "condition": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "event": {
            "type": "string"
          },
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "TriggerView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "enabled",
          "event",
          "action",
          "metadata",
          "version",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "condition": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "event": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  }
}